+++ src/utils/terminator.js
@@ -1,0 +1,12 @@
+const logger = require('../services/logger');
+
+module.exports = {
+  terminate(status, {
+    logs,
+  }) {
+    if (logs.length) {
+      logger.error(...logs);
+    }
+    process.exit(status);
+  },
+};