+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Added
 - Versioning - Configure the `warn-if-update-available` plugin that warns when a new version is available.
 - Prompt - Add the `@oclif/plugin-not-found` plugin which print "did you mean ...?" to correct the user input.
+- Authentication - Add Google authentication.
 
 ### Changed
 - Readme - Add a community section.